+++ pb_tool.cfg
@@ -43,7 +43,7 @@ extras: metadata.txt icon.png
 
 # Other directories to be deployed with the plugin.
 # These must be subdirectories under the plugin directory
-extra_dirs: api settings
+extra_dirs: api settings dmi_dialog_tabs
 
 # ISO code(s) for any locales (translations), separated by spaces.
 # Corresponding .ts files must exist in the i18n directory
